requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
python-decouple>=3.8
extract-msg>=0.44.0
python-dateutil>=2.8.0
pydantic>=2.0.0
typing-extensions>=4.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

